Add a validation check for elongations
At times when the elongation of Mercury or Venus crossed from a negative
to positive value, the check would mark this as a positive.

To avoid this, the elongation values are tested to see if they are
greater than five (as Mercury's greatest elongation is 18~28, and Venus'
is 45~47. If they are, then the check for an elongation can continue.