Update tests to match the restructure
Tests were updated to use the new modules, but were also simplified to
focus on testing value types rather than precise value data. Because a
lot of calculations resulted in floating point numbers or more, it's
more important to check that the properties are there (for now) rather
than whether the values are astronomically correct.