Redefine how transit times are stored
It was decided that an ordered list of transit times would function
better than a dictionary of key/value pairs. The dictionary was
converted to a list, with `get_transit_times` appending to a list in
a specific order of:

- `prev_body_rise` > `body_rise` | `body_set` > `next_body_rise`

A method, `format_transit_time`, was created to speed up the process of
appending rise times to the list, since each time would need to be
checked to see if it was a datetime value or a string, e.g. "NeverUp" or
"AlwaysUp".

Unit tests were updated to reflect this change as well.