Change the `event` key to `type` for events
The reason for this is so that any project that consumes the events
doesn't end up with a case like `event.event`. This is purely aesthetic,
but does keep in line with the transit times that use the key `type`
when describing if the transit is a rise or set.