Modify what transit times are returned
To ensure most transit times are covered for all latitudes and
longitudes, a total of four transit times are recorded in the array at
any one time. The pattern follows a rise/set/rise/set or
set/rise/set/rise, depending on if the object rises or sets first on the
given date.

The expectation is that libraries consuming this data will convert the
times from UTC to the required local time and filter out any times that
are not relevant.